drei/HTML 원리파악 및 로직수정
