+++ app/luckyshiba/page.tsx
@@ -6,6 +6,7 @@ import { Shiba } from '@/features/luckyshiba/components/shiba/Shiba';
 import TutorialOpener from '@/features/luckyshiba/components/tutorial';
 import CanvasLayout from '@/shared/components/3dmodel/Canvas';
 import Loading from '@/shared/components/Loading';
+import { ModalContextProvider } from '@/shared/components/portal/ModalContext';
 
 import { color } from '@/shared/contants/color';
 import { Physics } from '@react-three/cannon';
@@ -17,21 +18,23 @@ export default function LuckyShiba() {
     <div className="wrapper_3d">
       <Suspense fallback={<Loading />}>
         <SessionProvider>
-          <CanvasLayout color={color.bg} camera={{ position: [0, 2, 4] }}>
-            <Physics gravity={[0, -9.8, 0]}>
-              {/* <Debug> */}
-              <ambientLight />
-              <directionalLight position={[0, 5, 5]} />
-              <Shiba />
-              <Walls />
-              <Background />
-              {/* <DrawCall /> */}
-              {/* <StatsGl /> */}
+          <ModalContextProvider>
+            <CanvasLayout color={color.bg} camera={{ position: [0, 2, 4] }}>
+              <Physics gravity={[0, -9.8, 0]}>
+                {/* <Debug> */}
+                <ambientLight />
+                <directionalLight position={[0, 5, 5]} />
+                <Shiba />
+                <Walls />
+                <Background />
+                {/* <DrawCall /> */}
+                {/* <StatsGl /> */}
 
-              <TutorialOpener />
-              {/* </Debug> */}
-            </Physics>
-          </CanvasLayout>
+                <TutorialOpener />
+                {/* </Debug> */}
+              </Physics>
+            </CanvasLayout>
+          </ModalContextProvider>
         </SessionProvider>
       </Suspense>
     </div>
